+++ app/models/shift_reconciliation.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Numeric, Text
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -19,18 +19,18 @@ class ShiftReconciliation(Base):
     closed_by_user_id = Column(
         Integer, ForeignKey("users.id"), nullable=False, index=True
     )
-    cash_sales_total = Column(Float, nullable=False, default=0.0)
-    non_cash_sales_total = Column(Float, nullable=False, default=0.0)
-    refunds_total = Column(Float, nullable=False, default=0.0)
-    expected_cash = Column(Float, nullable=False, default=0.0)
-    counted_cash = Column(Float, nullable=False, default=0.0)
-    cash_variance = Column(Float, nullable=False, default=0.0)
-    expected_non_cash = Column(Float, nullable=False, default=0.0)
-    counted_non_cash = Column(Float, nullable=False, default=0.0)
-    non_cash_variance = Column(Float, nullable=False, default=0.0)
+    cash_sales_total = Column(Numeric(12, 2), nullable=False, default=0.0)
+    non_cash_sales_total = Column(Numeric(12, 2), nullable=False, default=0.0)
+    refunds_total = Column(Numeric(12, 2), nullable=False, default=0.0)
+    expected_cash = Column(Numeric(12, 2), nullable=False, default=0.0)
+    counted_cash = Column(Numeric(12, 2), nullable=False, default=0.0)
+    cash_variance = Column(Numeric(12, 2), nullable=False, default=0.0)
+    expected_non_cash = Column(Numeric(12, 2), nullable=False, default=0.0)
+    counted_non_cash = Column(Numeric(12, 2), nullable=False, default=0.0)
+    non_cash_variance = Column(Numeric(12, 2), nullable=False, default=0.0)
     completed_order_count = Column(Integer, nullable=False, default=0)
-    gross_sales_total = Column(Float, nullable=False, default=0.0)
-    net_sales_total = Column(Float, nullable=False, default=0.0)
+    gross_sales_total = Column(Numeric(12, 2), nullable=False, default=0.0)
+    net_sales_total = Column(Numeric(12, 2), nullable=False, default=0.0)
     notes = Column(Text, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
